+++ client/src/components/Commands/Saved.js
@@ -1,9 +1,6 @@
 // Include React as a dependency
 import React, { Component } from 'react'
 
-// Include the Helper (for the saved recall)
-import helpers from "../../utils/helpers"
-
 // Create the Main component
 class Saved extends Component {
