+++ client/src/components/Navbar.js
@@ -5,7 +5,7 @@ class Navbar extends Component {
 
     render() {
       return (
-        <div className="container">
+
             <nav className="navbar navbar-default">
               <div className="container-fluid">
                 <div className="navbar-header">
@@ -20,21 +20,19 @@ class Navbar extends Component {
                     <span className="icon-bar"></span>
                     <span className="icon-bar"></span>
                     </button>
-                    <Link className="navbar-brand" to="/">DigiFriend</Link>
+                    <Link className="navbar-brand" to={ (this.props.isLoggedIn ? "/dashboard" : "/") }>DigiFriend</Link>
                 </div>
 
                 <div className="collapse navbar-collapse navbar-ex1-collapse">
                     <ul className="nav navbar-nav navbar-left">
                     {/* Using <Link> in place of <a> and "to" in place of "href" */}
                     <li><Link to="/dashboard">Dashboard</Link></li>
-                    <li><Link to="/commands">Commands</Link></li>
                     <li><Link to="/settings">Settings</Link></li>
                     </ul>
                 </div>
 
               </div>
           </nav>
-        </div>
       );
     }
 };
